Add checks for 'spacesXYZ'
A change in 'colorSpec' of 'spacesXYZ' from imports to suggests and a test that triggers and error in the abscence of 'spacesXYZ' broke an example in 'ooacquire' and possibly in here. The checks added trigger only a warning and the functions return NA when 'spacesXYZ' is not available. This affects CRI, CCT and SSI.